getting ready to be used as scrawler component
