+++ composer.json
@@ -1,5 +1,5 @@
 {
-    "name": "ghost/route",
+    "name": "scrawler/router",
     "type": "library",
     "description": "An Fully Automatic RESTful PHP Router.",
     "keywords": [
@@ -9,7 +9,7 @@
         "restful",
         "automatic"
     ],
-    "homepage": "http://github.com/ghost-php/route",
+    "homepage": "http://github.com/scrawler-php/router",
     "license": "MIT",
     "authors": [
         {
@@ -18,12 +18,13 @@
             "homepage": "https://physcocode.com/",
             "role": "Developer"
         }
-        
+
     ],
     "require": {
-        "php": ">=5.3.0"
+      "symfony/http-foundation": "^4.0",
+      "symfony/http-kernel": "^4.0"
     },
    "autoload": {
-        "psr-4": {"Ghost\\Route\\": "src/"}
+        "psr-4": {"Scrawler\\Router\\": "src/"}
     }
 }
