+++ src/ControllerResolver.php
@@ -1,0 +1,15 @@
+<?php
+
+/**
+ * Argument Resolver Class
+ *
+ * @author : Pranjal Pandey
+ */
+
+namespace Scrawler\Router;
+use Symfony\Component\HttpKernel\Controller\ControllerResolver as BaseResolver;
+use Symfony\Component\HttpKernel\Controller\ControllerResolverInterface;
+
+class ConrollerResolver extends BaseResolver implements ControllerResolverInterface {
+
+}
