+++ src/HomeController.php
@@ -1,0 +1,39 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+
+
+class HomeController extends Controller
+{
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth');
+
+    }
+
+    /**
+     * Show the application dashboard.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function index(Request $r)
+    {
+      if(Auth::user()->first){
+        return redirect('/setup');
+      }
+        return view('feed');
+    }
+
+    public function logout(){
+        Auth::logout();
+        return redirect('');
+    }
+}
